Simplifying SidebarSidecarLayoutProps with new sidebarProps prop
