+++ src/components/sidebar/index.tsx
@@ -3,40 +3,8 @@ import useCurrentPath from 'hooks/use-current-path'
 import SidebarBackToLink from './components/sidebar-back-to-link'
 import SidebarFilterInput from './components/sidebar-filter-input'
 import SidebarNav from './components/sidebar-nav'
+import { MenuItem, SidebarProps } from './types'
 import s from './style.module.css'
-
-/**
- *
- * For reference: this is also defined in react-components/docs-sidenav:
- * https://github.com/hashicorp/react-components/blob/main/packages/docs-sidenav/types.ts
- *
- */
-export interface MenuItem {
-  divider?: boolean
-  fullPath?: string
-  hasActiveChild?: boolean
-  hasChildrenMatchingFilter?: boolean
-  href?: string
-  id?: string
-  isActive?: boolean
-  matchesFilter?: boolean
-  path?: string
-  routes?: MenuItem[]
-  title?: string
-  /* Temporary solution to allow rendering of unlinked headings, as in designs */
-  heading?: string
-}
-
-interface SidebarProps {
-  menuItems: MenuItem[]
-  /** title to be shown as the title of the sidebar */
-  title: string
-  /** Optional { text, url } to use for the "← Back to..." link at the top of the sidebar */
-  backToLink?: {
-    text: string
-    url: string
-  }
-}
 
 const addItemMetadata = (
   currentPath: string,
@@ -74,7 +42,7 @@ const addItemMetadata = (
 /**
  * This does not use Array.filter because we need to add metadata to each item
  * that is used for determining the open/closed state of submenu items.
- * */
+ */
 const getFilteredMenuItems = (items: MenuItem[], filterValue: string) => {
   if (!filterValue) {
     return items
@@ -121,8 +89,8 @@ const getFilteredMenuItems = (items: MenuItem[], filterValue: string) => {
 }
 
 const Sidebar: React.FC<SidebarProps> = ({
-  menuItems,
   backToLink = {},
+  menuItems,
   title,
 }) => {
   const currentPath = useCurrentPath({ excludeHash: true, excludeSearch: true })
@@ -143,4 +111,5 @@ const Sidebar: React.FC<SidebarProps> = ({
   )
 }
 
+export type { MenuItem, SidebarProps }
 export default Sidebar
