+++ src/layouts/sidebar-sidecar/server.ts
@@ -91,15 +91,17 @@ export function getStaticGenerationFunctions({
 
       const finalProps = {
         layoutProps: {
-          backToLink: {
-            text: `Back to ${product.name}`,
-            url: `/${product.slug}`,
-          },
           breadcrumbLinks,
           githubFileUrl,
           headings,
-          navData: navDataWithFullPaths,
-          productName: product.name,
+          sidebarProps: {
+            backToLink: {
+              text: `Back to ${product.name}`,
+              url: `/${product.slug}`,
+            },
+            menuItems: navDataWithFullPaths,
+            title: product.name,
+          },
         },
         mdxSource,
         product,
